Push changes!
* Initial work on WIREDs.
* Added MessengerInviteFriends.
* Some misc changes.

+++ server/src/main/kotlin/tk/jomp16/habbo/communication/incoming/room/RoomItemUseClothingHandler.kt
@@ -38,13 +38,19 @@ class RoomItemUseClothingHandler {
 
         if (roomItem.userId != habboSession.userInformation.id || !roomItem.furnishing.itemName.startsWith("clothing_")) return
 
+        if (habboSession.userInformation.clothings.contains(roomItem.itemName)) {
+            habboSession.sendSuperNotification("", "title", "\${notification.figureset.already.redeemed.title}", "message", "\${notification.figureset.already.redeemed.message}", "linkUrl", "\${notification.figureset.already.redeemed.linkUrl}", "linkTitle", "\${notification.figureset.already.redeemed.linkTitle}")
+
+            return
+        }
+
         habboSession.currentRoom!!.removeItem(habboSession.roomUser!!, roomItem)
 
         habboSession.userInformation.clothings += roomItem.itemName
 
         habboSession.sendHabboResponse(Outgoing.USER_CLOTHINGS, habboSession.userInformation.clothings)
 
-        habboSession.sendSuperNotification("", "title", "\${notification.figureset.redeemed.success.title}", "message", "\${notification.figureset.redeemed.success.message}", "linkUrl", "event:avatareditor/open", "linkTitle", "\${notification.figureset.redeemed.success.linkTitle}")
+        habboSession.sendSuperNotification("", "title", "\${notification.figureset.redeemed.success.title}", "message", "\${notification.figureset.redeemed.success.message}", "linkUrl", "\${notification.figureset.redeemed.success.linkUrl}", "linkTitle", "\${notification.figureset.redeemed.success.linkTitle}")
 
         ClothingDao.addClothing(habboSession.userInformation.id, roomItem.itemName)
         ItemDao.deleteItem(itemId)
